+++ rudof_lib/src/api/data/mod.rs
@@ -3,3 +3,4 @@ mod data_operations_trait;
 pub mod implementations;
 
 pub use data_operations_trait::DataOperations;
+pub(crate) use implementations::write_pretty_json;